requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
asgiref==3.2.10           # via django
backcall==0.2.0           # via ipython
click==7.1.2              # via pip-tools
decorator==4.4.2          # via ipython, traitlets
django-environ==0.4.5     # via -r requirements.in
django-log-request-id==1.5.0  # via -r requirements.in
django==3.1               # via -r requirements.in, django-log-request-id
gunicorn==20.0.4          # via -r requirements.in
honcho==1.0.1             # via -r requirements.in
ipython-genutils==0.2.0   # via traitlets
ipython==7.16.1           # via -r requirements.in
jedi==0.17.1              # via ipython
parso==0.7.0              # via jedi
pendulum==2.1.2           # via -r requirements.in
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
pillow==7.2.0             # via -r requirements.in
pip-tools==5.2.1          # via -r requirements.in
prompt-toolkit==3.0.5     # via ipython
psycopg2==2.8.5           # via -r requirements.in
ptyprocess==0.6.0         # via pexpect
pygments==2.6.1           # via ipython
python-dateutil==2.8.1    # via pendulum
pytz==2020.1              # via django
pytzdata==2020.1          # via pendulum
six==1.15.0               # via pip-tools, python-dateutil, traitlets
sqlparse==0.3.1           # via django
traitlets==4.3.3          # via ipython
wcwidth==0.2.5            # via prompt-toolkit
whitenoise==5.1.0         # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools